+++ test/libsolidity/semanticTests/functionCall/conditional_with_arguments.sol
@@ -1,0 +1,10 @@
+contract C {
+    function g(int x, int y) public pure returns (int) { return x - y; }
+    function h(int y, int x) public pure returns (int) { return y - x; }
+
+    function f() public pure returns (int) {
+        return (true ? g : h)(1, 2);
+    }
+}
+// ----
+// f() -> -1